lifeart-sales-auto/scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
lxml>=4.9.0
